[bug/enhancement] Multiple fixes / extra formatting (#24)
* Fix borders in circle.py 

The border_width and border_color variables were initialized but never used. This made it so there was no possibility to use borders. I fixed this issue by first drawing the border and after that drawing the inner parts (width/len - border). 

Signed-off-by: T <thollllllllllllllll@gmail.com>

* fix borders in box.py and add border radius (rounding) feature

The code fixes the same issue that occured in circle.py 
e.g it sets a self._border_width var to the given value but doesnt do anything with it. The new code draws the border and the inner rectangle on top of it. 

Also a new feature added is the usage of a new var (self._border_radius), since pygame.draw.rect supports rounding a rectangle with a given radius it wasn't too hard to add. 

Signed-off-by: T <thollllllllllllllll@gmail.com>

* Add border_radius param to generators.py

Adds the border_radius param to generators.py for use in play.new_box()

Signed-off-by: T <thollllllllllllllll@gmail.com>

* Update generators.py (formatting)

- Made a more readable format for intellisense (vscode) / other documentation.

- Added error for if words variable is not a string (instead of unicode error)
 

Signed-off-by: T <thollllllllllllllll@gmail.com>

* Update generators.py

Added extra documentation

Signed-off-by: T <thollllllllllllllll@gmail.com>

* Update image.py

Added more support for loading images

Signed-off-by: T <thollllllllllllllll@gmail.com>

* Add font loading function to text.py 

Added a font loading function to text.py so it can load font files (.ttf) and otherwise will stick to the system default.

Signed-off-by: T <thollllllllllllllll@gmail.com>

* fix: returning too general type and improvements to default arguments

Signed-off-by: Koen <98043234+koen1711@users.noreply.github.com>

* Update setup.py

incremented minor version number

Signed-off-by: T <thollllllllllllllll@gmail.com>

* Update text.py

Update error logging

Signed-off-by: T <thollllllllllllllll@gmail.com>

* Use black ./

Used black ./ to format all the changes files

* pylint test and changes

Did a pylint check and fixed/changed some code to eventually make the current score 9.89/10

* Update text.py for pylint check

fixed pylint error

- broad-exception-caught
- inconsistent-return-statements

* Fix / add physics to box / circle

fixes and adds logic for rotating the boxes / circles which there was no code for.

* formatted with black

---------

Signed-off-by: T <thollllllllllllllll@gmail.com>
Signed-off-by: Koen <98043234+koen1711@users.noreply.github.com>
Co-authored-by: Koen <98043234+koen1711@users.noreply.github.com>